Implements audio output and comments code
Implements playing audio with the Port Audio 19 library, using a
callback function.

Currently this is not asynchronous, the audio data is passed in when the
audio stream is opened, and includes the unfiltered and filtered audio,
separated by 1 second of silence (samplerate samples).

Also adds comments to all functions, and describes what is going on in
each major section of main.